fine-tune/requirements.txt
transformers==4.57.6
accelerate==1.12.0
kernels
torch==2.5.1
torchvision==0.20.1
peft==0.18.0
bitsandbytes==0.44.0
datasets>=2.18.0
pillow>=10.0.0
pandas>=2.0.0
tqdm>=4.66.0
